Restructure controllers by feature/function instead of roles
- Moved authentication functions to Authentication/ folder
- Moved user profile functions to ProfileUser/ folder
- Moved seller profile functions to ProfileSeller/ folder
- Moved admin user management to AdminUserManagement/ folder
- Moved admin category management to AdminCategoryManagement/ folder
- Removed old role-based controller files
- Maintained all existing API routes and functionality
