+++ PlatformFlower/Controllers/AdminUserManagement/UserStatusController.cs
@@ -7,19 +7,19 @@ using PlatformFlower.Services.Common.Logging;
 using PlatformFlower.Services.Common.Response;
 using PlatformFlower.Services.Common.Validation;
 
-namespace PlatformFlower.Controllers
+namespace PlatformFlower.Controllers.AdminUserManagement
 {
     [ApiController]
-    [Route("api/[controller]")]
+    [Route("api/admin")]
     [Authorize(Roles = "admin")]
-    public class AdminController : ControllerBase
+    public class UserStatusController : ControllerBase
     {
         private readonly IUserManagementService _userManagementService;
         private readonly IResponseService _responseService;
         private readonly IValidationService _validationService;
         private readonly IAppLogger _logger;
 
-        public AdminController(
+        public UserStatusController(
             IUserManagementService userManagementService,
             IResponseService responseService,
             IValidationService validationService,
@@ -29,64 +29,6 @@ namespace PlatformFlower.Controllers
             _responseService = responseService;
             _validationService = validationService;
             _logger = logger;
-        }
-
-        [HttpGet("users")]
-        public async Task<ActionResult<ApiResponse<PaginatedUsersResponseDto>>> GetUsers(
-            [FromQuery] int pageNumber = 1,
-            [FromQuery] int pageSize = 10,
-            [FromQuery] string? searchTerm = null,
-            [FromQuery] string? userType = null,
-            [FromQuery] bool? isActive = null)
-        {
-            try
-            {
-                _logger.LogInformation($"Admin getting users list - Page: {pageNumber}, Size: {pageSize}");
-
-                if (pageNumber < 1) pageNumber = 1;
-                if (pageSize < 1 || pageSize > 100) pageSize = 10;
-
-                var result = await _userManagementService.GetUsersAsync(pageNumber, pageSize, searchTerm, userType, isActive);
-
-                var response = _responseService.CreateSuccessResponse(result, "Users retrieved successfully");
-                return Ok(response);
-            }
-            catch (Exception ex)
-            {
-                _logger.LogError($"Error getting users list: {ex.Message}", ex);
-                var response = _responseService.CreateErrorResponse<PaginatedUsersResponseDto>(
-                    "An error occurred while retrieving users"
-                );
-                return StatusCode(500, response);
-            }
-        }
-
-        [HttpGet("users/{id}")]
-        public async Task<ActionResult<ApiResponse<UserDetailResponseDto>>> GetUserById(int id)
-        {
-            try
-            {
-                _logger.LogInformation($"Admin getting user details for ID: {id}");
-
-                var user = await _userManagementService.GetUserByIdAsync(id);
-                
-                if (user == null)
-                {
-                    var notFoundResponse = _responseService.CreateErrorResponse<UserDetailResponseDto>("User not found");
-                    return NotFound(notFoundResponse);
-                }
-
-                var response = _responseService.CreateSuccessResponse(user, "User details retrieved successfully");
-                return Ok(response);
-            }
-            catch (Exception ex)
-            {
-                _logger.LogError($"Error getting user details for ID {id}: {ex.Message}", ex);
-                var response = _responseService.CreateErrorResponse<UserDetailResponseDto>(
-                    "An error occurred while retrieving user details"
-                );
-                return StatusCode(500, response);
-            }
         }
 
         [HttpPut("users/{id}/toggle-status")]
